chore: Extract text-colouring to doctest/parts/color.h
